# docs/swagger.yaml
basePath: /
definitions:
  schemas.AllCollections:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  schemas.AllCollectionsResp:
    properties:
      collections:
        items:
          $ref: '#/definitions/schemas.AllCollections'
        type: array
    type: object
...
  schemas.CreateCollectionReq:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  schemas.CreateCollectionResp:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  schemas.CreateUserReq:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  schemas.CreateUserResp:
    properties:
      id:
        type: integer
      token:
        type: string
      username:
        type: string
    type: object
...
  schemas.GetCollectionByIDResp:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  schemas.GetProfileResp:
    properties:
      profile:
        $ref: '#/definitions/schemas.Profile'
    type: object
  schemas.PatchCollectionReq:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    required:
    - id
    type: object
  schemas.Profile:
    properties:
      id:
        type: integer
      username:
        type: string
    type: object
  schemas.SignInReq:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  schemas.SignInResp:
    properties:
      id:
        type: integer
      token:
        type: string
      username:
        type: string
    type: object
...
  schemas.UpdateCardReq:
    properties:
      answer:
        type: string
      id:
        type: integer
      question:
        type: string
    required:
    - id
    type: object
  schemas.UpdateCardResp:
    properties:
      answer:
        type: string
      collection_id:
        type: integer
      id:
        type: integer
      question:
        type: string
    type: object
  schemas.UpdateCollectionReq:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    required:
    - id
    - name
    type: object
  schemas.UpdateCollectionResp:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  schemas.UpdatePasswordReq:
    properties:
      id:
        type: integer
      new_password:
        type: string
      old_password:
        type: string
    required:
    - id
    - new_password
    - old_password
    type: object
  schemas.UpdateUsernameReq:
    properties:
      id:
        type: integer
      username:
        type: string
    required:
    - id
    - username
    type: object
  schemas.UpdateUsernameResp:
    properties:
      id:
        type: integer
      username:
        type: string
    type: object
  util.ErrorResponse:
    properties:
      error:
        type: string
    type: object
info:
  contact: {}
  description: REST API for flashcard learning platform.
  title: Card Backend API
  version: "1.1"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: signin payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.SignInReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.SignInResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      summary: Login user
      tags:
      - auth
...
  /auth/signup:
    post:
      consumes:
      - application/json
      parameters:
      - description: signup payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateUserReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/schemas.CreateUserResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      summary: Register user
      tags:
      - auth
  /cards/{cardID}/:
    delete:
      parameters:
      - description: card id
        in: path
        name: cardID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete card
      tags:
      - cards
    patch:
      consumes:
      - application/json
      parameters:
      - description: card id
        in: path
        name: cardID
        required: true
        type: integer
      - description: card payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateCardReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.UpdateCardResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update card
      tags:
      - cards
    put:
      consumes:
      - application/json
      parameters:
      - description: card id
        in: path
        name: cardID
        required: true
        type: integer
      - description: card payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateCardReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.UpdateCardResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update card
      tags:
      - cards
  /collections/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.AllCollectionsResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all collections
      tags:
      - collections
    post:
      consumes:
      - application/json
      parameters:
      - description: collection payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.CreateCollectionReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/schemas.CreateCollectionResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create collection
      tags:
      - collections
  /collections/{collectionID}/:
    delete:
      parameters:
      - description: collection id
        in: path
        name: collectionID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete collection
      tags:
      - collections
    get:
      parameters:
      - description: collection id
        in: path
        name: collectionID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.GetCollectionByIDResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get collection by ID
      tags:
      - collections
    patch:
      consumes:
      - application/json
      parameters:
      - description: collection id
        in: path
        name: collectionID
        required: true
        type: integer
      - description: collection payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.PatchCollectionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.UpdateCollectionResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Partial update collection
      tags:
      - collections
    put:
      consumes:
      - application/json
      parameters:
      - description: collection id
        in: path
        name: collectionID
        required: true
        type: integer
      - description: collection payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateCollectionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.UpdateCollectionResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Full update collection
      tags:
      - collections
...
  /profile/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.GetProfileResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get profile
      tags:
      - profile
  /profile/password:
    put:
      consumes:
      - application/json
      parameters:
      - description: password payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdatePasswordReq'
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update password
      tags:
      - profile
  /profile/username:
    put:
      consumes:
      - application/json
      parameters:
      - description: username payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.UpdateUsernameReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.UpdateUsernameResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update username
      tags:
      - profile
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"